Simplify URLOpener abstraction and model behaviour
