+++ Domain Model/EurofurenceModel/Public/Builder/Configurable Dependencies/URLOpener.swift
@@ -2,7 +2,6 @@ import Foundation
 
 public protocol URLOpener {
 
-    func canOpen(_ url: URL) -> Bool
     func open(_ url: URL)
 
 }
