+++ Eurofurence/Application Adapter Implementations/AppURLOpener.swift
@@ -1,12 +1,8 @@
 import EurofurenceModel
-import Foundation.NSURL
-import UIKit.UIApplication
+import Foundation
+import UIKit
 
 struct AppURLOpener: URLOpener {
-
-    func canOpen(_ url: URL) -> Bool {
-        return UIApplication.shared.canOpenURL(url)
-    }
 
     func open(_ url: URL) {
         UIApplication.shared.open(url)
